more messingwith CAN

+++ lib/MathOps/src/MathOps.cpp
@@ -1,4 +1,5 @@
 #include "MathOps.hpp"
+#include <stdint.h>
 
 
 float fmaxf(float x, float y){
@@ -35,7 +36,7 @@ int float_to_uint(float x, float x_min, float x_max, int bits){
     /// Converts a float to an unsigned int, given range and number of bits ///
     float span = x_max - x_min;
     float offset = x_min;
-    return (int) ((x-offset)*((float)((1<<bits)-1))/span);
+    return (int) ((x-offset)*((float)(((int32_t)(1)<<bits)-1))/span);
     }
     
     
@@ -43,5 +44,5 @@ float uint_to_float(int x_int, float x_min, float x_max, int bits){
     /// converts unsigned int to float, given range and number of bits ///
     float span = x_max - x_min;
     float offset = x_min;
-    return ((float)x_int)*span/((float)((1<<bits)-1)) + offset;
+    return ((float)x_int)*span/((float)(((int32_t)(1)<<bits)-1)) + offset;
     }
